% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{annotate_cluster}
\alias{annotate_cluster}
\title{Annotate a single cluster}
\usage{
annotate_cluster(
  identifying_data,
  cluster_id,
  couple_ids = NULL,
  compound_levels = NULL,
  biosoc_levels = NULL
)
}
\arguments{
\item{identifying_data}{A data.frame containing possible identifications.}

\item{cluster_id}{A string indicating the cluster id.}

\item{couple_ids}{A 2-tuple indicating specific compounds to consider for annotation}

\item{compound_levels}{A vector containing the compound levels to consider, in the given order.}

\item{biosoc_levels}{A vector containing the biosource levels to consider, in the given order.}
}
\value{
A data.frame with the selected identification annotated.
}
\description{
Annotate a single cluster
}
