% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_architecture_for_keep_top_peaks}
\alias{check_architecture_for_keep_top_peaks}
\title{Check project architecture before running \code{\link{keep_top_peaks}}.}
\usage{
check_architecture_for_keep_top_peaks(export_filtered_peaks)
}
\arguments{
\item{export_filtered_peaks}{Copy peaks files corresponding to the peaks remaining after filtering in a new folder.}
}
\description{
Check project architecture before running \code{\link{keep_top_peaks}}.
}
