% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_boolean}
\alias{check_boolean}
\title{Check if value is boolean.}
\usage{
check_boolean(x, n)
}
\arguments{
\item{x}{Value to check.}

\item{n}{Name of value.}
}
\description{
Check if value is boolean.
}
