% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_for_convert_csv_to_msp}
\alias{check_for_convert_csv_to_msp}
\title{Check before running \code{\link{launch_msfinder_annotation}}.}
\usage{
check_for_convert_csv_to_msp(min_score)
}
\arguments{
\item{min_score}{Input parameter to check.}
}
\description{
Check before running \code{\link{launch_msfinder_annotation}}.
}
