% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_input_parameters_launch_msfinder}
\alias{check_input_parameters_launch_msfinder}
\title{Check if input parameters are correct before running \code{launch_msfinder_annotation}.}
\usage{
check_input_parameters_launch_msfinder(msfinder_biosoc_levels,
  msfinder_compound_levels, levels_scores, overwrite)
}
\arguments{
\item{msfinder_biosoc_levels, msfinder_compound_levels, levels_scores, overwrite}{Input parameters to check.}
}
\description{
Check if input parameters are correct before running \code{launch_msfinder_annotation}.
}
