% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_main_architecture}
\alias{check_main_architecture}
\title{Check the project main architecture}
\usage{
check_main_architecture(posneg_check = TRUE)
}
\arguments{
\item{posneg_check}{Whether to check the existence of pos/neg directories.}
}
\description{
Check the project main architecture
}
