% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_path}
\alias{check_path}
\title{Check if a path exist.}
\usage{
check_path(filetype, stop = TRUE, na_msg = NA, ...)
}
\arguments{
\item{filetype}{A string indicating the needed directory or file.}

\item{stop}{If \code{TRUE}, stops the script if a path is not found.}

\item{na_msg}{A string with info to print if path doesn't exist.}

\item{...}{Additional parameters passed on to \code{\link{get_project_file_path}}.}
}
\description{
Check if a path exist.
}
