% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_positive_int}
\alias{check_positive_int}
\title{Check if value is a positive integer.}
\usage{
check_positive_int(x, n)
}
\arguments{
\item{x}{Value to check.}

\item{n}{Name of value.}
}
\description{
Check if value is a positive integer.
}
