% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_positive_num}
\alias{check_positive_num}
\title{Check if value is a positive numerical value.}
\usage{
check_positive_num(x, n, no_NA = TRUE)
}
\arguments{
\item{x}{Value to check.}

\item{n}{Name of value.}

\item{no_NA}{if TRUE, \code{x} can't be NA.}
}
\description{
Check if value is a positive numerical value.
}
