% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_probability}
\alias{check_probability}
\title{Check if value is a probability (numerical value between 0 and 1).}
\usage{
check_probability(x, n)
}
\arguments{
\item{x}{Value to check.}

\item{n}{Name of value.}
}
\description{
Check if value is a probability (numerical value between 0 and 1).
}
