% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_references}
\alias{check_references}
\title{Check if value is a correct reference data.frame}
\usage{
check_references(x, n)
}
\arguments{
\item{x}{Value to check.}

\item{n}{Name of value.}
}
\description{
Check if value is a correct reference data.frame
}
