% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_parameters.R
\name{choose_project_directory}
\alias{choose_project_directory}
\title{Declare the project directory for the analysis process.}
\usage{
choose_project_directory(project_directory, overwrite = FALSE)
}
\arguments{
\item{project_directory}{The path of the project directory.}

\item{overwrite}{A boolean indicating whether to stop the script if there is an existing analysis or overwrite it.}
}
\description{
Declare the project directory for the analysis process.
}
