% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cad_pda_functions.R
\name{clean_cad_pda_peaks}
\alias{clean_cad_pda_peaks}
\title{Clean peaks from CAD/PDA peaks based on manually extracted data..}
\usage{
clean_cad_pda_peaks(project_directory, mode, data_start_row = 5,
  plot_spectra = FALSE, plot_ROI = FALSE)
}
\arguments{
\item{project_directory}{The path of the project directory.}

\item{mode}{"CAD" or "PDA".}

\item{data_start_row}{An integer indicating on which row of the manually extracted file starts the actual data (retention times and intensities).}

\item{plot_spectra}{A boolean indicating whether to plot the spectra or not.}

\item{plot_ROI}{A boolean indicating whether to plot the ROI or not.}
}
\description{
The file containing CAD/PDA information manually extracted must have the following format: a 2n-columns data.frame representing CAD/PDA information manually extracted, with each column alternatively containing retention times and intensity.
}
