% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{clean_for_current_run}
\alias{clean_for_current_run}
\title{Create a directory or delete a file for a new run.}
\usage{
clean_for_current_run(filetype, delete = TRUE, create_dir = FALSE, ...)
}
\arguments{
\item{filetype}{A string indicating the needed directory or file.}

\item{delete}{A boolean indicating whether the file or directory indicated by the path needs to be deleted.}

\item{create_dir}{A boolean indicating whether the directory indicated by the path needs to be created.}

\item{...}{Additional parameters to pass to \code{\link{get_project_file_path}}.}
}
\description{
Rename the old directory or file if existing.
}
