% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts_nl_functions.R
\name{clean_msc_combinations}
\alias{clean_msc_combinations}
\title{Cleans the columns name of data generated my MSCombine.}
\usage{
clean_msc_combinations(data_to_clean)
}
\arguments{
\item{data_to_clean}{A data.frame with columns to rename.}
}
\value{
A data.frame containing the same data than \code{data_to_clean} but with new column names.
}
\description{
Cleans the columns name of data generated my MSCombine.
}
