% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdial_functions.R
\name{clear_duplicate_links}
\alias{clear_duplicate_links}
\title{Clear duplicate undirected MSDial links.}
\usage{
clear_duplicate_links(links)
}
\arguments{
\item{links}{A 6-columns data.frame of links: id of 1st peak, id of 2nd peak, nature, adduct ion name for 1st peak, adduct ion name for 2nd peak, correlation.}
}
\value{
\code{links} without duplicate links.
}
\description{
Clear duplicate undirected MSDial links.
}
