% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts_nl_functions.R
\name{compute_massdiff}
\alias{compute_massdiff}
\title{Compute mass differences between 2 lists of adducts}
\usage{
compute_massdiff(
  user_adduct_pos = NA,
  user_adduct_neg = NA,
  significance_mass_threshold = NA
)
}
\arguments{
\item{user_adduct_pos, user_adduct_neg}{An optional 2-column data.frame containing information about positive adducts or negative adducts (one column for the name and one column for the mass difference with the base compound). If no data.frame is provided, the package default list is used.}

\item{significance_mass_threshold}{A numeric value indicating at which threshold mass differences are discarded because too small.}
}
\value{
A 3-columns data.frame with the pairs of adducts and their mass difference values.
}
\description{
Compute mass differences between 2 lists of adducts
}
