% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{concat_msp_files}
\alias{concat_msp_files}
\title{Concatenate all MSP files present in the folder at \code{path}.}
\usage{
concat_msp_files(path, output_file = "_all_.msp")
}
\arguments{
\item{path}{The path of the folder containing MSP files to concatenate.}

\item{output_file}{The name of the concatenated file.}
}
\description{
Concatenate all MSP files present in the folder at \code{path}.
}
