% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_csv_to_msp.R
\name{convert_csv_to_msp}
\alias{convert_csv_to_msp}
\title{Convert the final CSV file post annotations to MSP format.}
\usage{
convert_csv_to_msp(all = FALSE, min_score = 20)
}
\arguments{
\item{all}{All peaks are exported to the MSP files.}

\item{min_score}{If \code{all} is \code{FALSE}, peaks must have a final score >= \code{min_score} to be exported to the MSP files.}
}
\description{
Convert the final CSV file post annotations to MSP format.
}
