% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts_nl_functions.R
\name{detect_posneg_combinations}
\alias{detect_posneg_combinations}
\title{Detect possible adducts, neutral losses and positive and negative peaks to combine.}
\usage{
detect_posneg_combinations(peaks_data_pos, peaks_data_neg, threshold_rt,
  threshold_mz, adducts_massdiff)
}
\arguments{
\item{peaks_data_pos}{A data.frame containing positive peaks to treat, coming from MSDial.}

\item{peaks_data_neg}{A data.frame containing negative peaks to treat, coming from MSDial.}

\item{threshold_rt}{A numerical value indicating the retention time tolerance.}

\item{threshold_mz}{A numerical value indicating the mass tolerance in Dalton for the detection of adducts and neutral losses.}

\item{adducts_massdiff}{A data.frame containing combinations information (adduct1 adduct2 m).}
}
\value{
A data.frame of possible links between positive and negative peaks.
}
\description{
Detect possible adducts, neutral losses and positive and negative peaks to combine.
}
