% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts_nl_functions.R
\name{detect_single_polarity_combinations}
\alias{detect_single_polarity_combinations}
\title{Detect possible adducts and neutral losses.}
\usage{
detect_single_polarity_combinations(
  peaks_data,
  threshold_rt,
  threshold_mz,
  mass_neutral_loss,
  mass_isotopes
)
}
\arguments{
\item{peaks_data}{A data.frame containing peaks to treat, coming from MSDial.}

\item{threshold_rt}{A numerical value indicating the retention time tolerance.}

\item{threshold_mz}{A numerical value indicating the mass tolerance in Dalton for the detection of adducts and neutral losses.}

\item{mass_neutral_loss}{An 2-column data.frame containing information about neutral losses (one column for the name and one column for the mass difference with the base compound).}

\item{mass_isotopes}{An 2-column data.frame containing information about isotopes (one column for the name and one column for the mass difference with the base compound).}
}
\value{
A data.frame of possible links due to neutral losses and isotopes.
}
\description{
Detect possible adducts and neutral losses.
}
