% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_outputs.R
\name{export_data}
\alias{export_data}
\title{Export data for users.}
\usage{
export_data(data_to_export, project_directory, filetype, source = NA,
  empty_na = FALSE)
}
\arguments{
\item{data_to_export}{A data.frame to export to the user file system.}

\item{project_directory}{The path of the project directory.}

\item{filetype}{A string indicating the needed directory or file.}

\item{source}{A string indicating which mode to read from: "pos" or "neg".}

\item{empty_na}{A boolean indicating if NA must be replaced by an empty string in the output file.}
}
\description{
Export data for users.
}
