% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extract_concatenated_data_from_column}
\alias{extract_concatenated_data_from_column}
\title{Extract data concatenated in a single column.}
\usage{
extract_concatenated_data_from_column(column_to_divide)
}
\arguments{
\item{column_to_divide}{A 1-column data.frame with data to divide in several columns, with format "key1=value1,key2=value2,...".}
}
\value{
A data.frame containing X columns.
}
\description{
Extract data concatenated in a single column.
}
