% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdial_functions.R
\name{extract_msdial_links}
\alias{extract_msdial_links}
\title{Extract links between elements identified by MSDial.}
\usage{
extract_msdial_links(msdial_data)
}
\arguments{
\item{msdial_data}{A 4-columns data.frame with for each peak the alignment id, the global id, the adduct ion name (column "Adduct.type" by default) and the links detected by MSDial (column "Post.curation.result" by default).}
}
\value{
A 6-columns data.frame of links: id of 1st peak, id of 2nd peak, nature, adduct ion name for 1st peak, adduct ion name for 2nd peak, correlation.
}
\description{
Extract links between elements identified by MSDial.
}
