% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{find_structure}
\alias{find_structure}
\title{Find the most probable structure by iterating over all possible levels of annotations.}
\usage{
find_structure(possibilities, compound_levels = NULL, biosoc_levels = NULL)
}
\arguments{
\item{possibilities}{A data.frame of possible structures.}

\item{compound_levels}{A vector containing the compound levels to consider, in the given order.}

\item{biosoc_levels}{A vector containing the biosource levels to consider, in the given order.}
}
\value{
A vector containing the row id(s) of the most probable structure(s).
}
\description{
Find the most probable structure by iterating over all possible levels of annotations.
}
