% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adducts_nl_functions.R
\name{get_adducts_nl_links}
\alias{get_adducts_nl_links}
\title{Detect possible neutral losses, isotopes and adduct links.}
\usage{
get_adducts_nl_links(data_to_treat, samples_cols, threshold_rt,
  threshold_mz, detected_adducts, user_neutral_refs = NA,
  user_pos_adducts_refs = NA, user_neg_adducts_refs = NA)
}
\arguments{
\item{data_to_treat}{A data.frame containing peaks data.}

\item{samples_cols}{A list of string containing the names of samples columns.}

\item{threshold_rt}{A numerical value indicating the retention time tolerance.}

\item{threshold_mz}{A numerical value indicating the mass tolerance in Dalton for the detection of adducts and neutral losses.}

\item{detected_adducts}{A data.frame indicating which adducts were detected by MSDial.}

\item{user_neutral_refs, user_pos_adducts_refs, user_neg_adducts_refs}{An optional 2-column data.frame containing information about neutral losses, positive adducts or negative adducts (one column for the name and one column for the mass difference with the base compound). If no data.frame is provided, the package default list is used.}
}
\value{
A list of 3 data.frames: "links" containing detected links, "computed_massdiff" containing all pairs of positive/negative adducts pair mass differences and "filtered_massdiff" containing pairs of positive/negative adducts pair mass differences for adducts pair detected by MSDial.
}
\description{
Detect possible neutral losses, isotopes and adduct links.
}
