% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_confirmation_for_overwriting}
\alias{get_confirmation_for_overwriting}
\title{Ask the user confirmation for overwriting if the path already exists.}
\usage{
get_confirmation_for_overwriting(path)
}
\arguments{
\item{path}{A string indicating the path meant to be overwritten.}
}
\description{
Ask the user confirmation for overwriting if the path already exists.
}
