% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{get_most_probable}
\alias{get_most_probable}
\title{Get the most probable structure for the whole input data.frame.}
\usage{
get_most_probable(possibilities)
}
\arguments{
\item{possibilities}{A data.frame of possible structures.}
}
\value{
A vector containing the row id(s) of the most probable structure(s).
}
\description{
Get the most probable structure for the whole input data.frame.
}
