% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_outputs.R
\name{get_pattern_file}
\alias{get_pattern_file}
\title{Get path for a file found thanks to a pattern.}
\usage{
get_pattern_file(dir_path, pattern, only_name = FALSE)
}
\arguments{
\item{dir_path}{A file.path indicating the directory in which to look for the file.}

\item{pattern}{A string indicating the pattern used to find the file.}

\item{only_name}{A boolean indicating whether to return the full path or only the name of the file.}
}
\value{
A string with the found file path or NA if not found.
}
\description{
Get path for a file found thanks to a pattern.
}
