% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_outputs.R
\name{get_project_file_path}
\alias{get_project_file_path}
\title{Get path for a project file or directory.}
\usage{
get_project_file_path(
  filetype,
  source = NA,
  msfinder_info = NA,
  msfinder_lvl = NA,
  peak_id = NA
)
}
\arguments{
\item{filetype}{A string indicating the needed directory or file.}

\item{source}{A string indicating which mode to read from: "pos" or "neg".}

\item{msfinder_info}{A string indicating which type of MSFinder file to get path from: "Structure" or "Formula".}

\item{msfinder_lvl}{A string indicating which level of MSFinder file to get path from.}

\item{peak_id}{A string indicating the wanted peak id.}
}
\value{
The file or folder full path.
}
\description{
Get path for a project file or directory.
}
