% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdial_functions.R
\name{get_samples_info}
\alias{get_samples_info}
\title{Get samples names and class from an MSDial file and generate column names.}
\usage{
get_samples_info(source = "pos")
}
\arguments{
\item{source}{A string indicating which mode to read from: "pos" or "neg".}
}
\value{
A data.frame with the samples names, their class, filetype and corresponding column name.
}
\description{
Get samples names and class from an MSDial file and generate column names.
}
