% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_outputs.R
\name{import_data}
\alias{import_data}
\title{Import data exported by the script.}
\usage{
import_data(project_directory, filetype, ...)
}
\arguments{
\item{project_directory}{The path of the project directory.}

\item{filetype}{A string indicating the needed directory or file.}

\item{...}{Additional parameters passed on to \code{\link{get_project_file_path}}.}
}
\description{
Import data exported by the script.
}
