% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdial_functions.R
\name{import_msdial_data}
\alias{import_msdial_data}
\title{Import all data generated by MSDial.}
\usage{
import_msdial_data(project_directory, filter_blk, filter_blk_threshold,
  filter_mz, filter_rsd, filter_rsd_threshold)
}
\arguments{
\item{project_directory}{The path of the project directory.}

\item{filter_blk}{A boolean indicating whether or not to delete rows with too much noise.}

\item{filter_blk_threshold}{A numerical threshold for noise filtering: rows with ratio mean(blank columns)/mean(qc columns) >= \code{filter_blk} are deleted (if there are no QC columns in the sample, the mean of standard columns is used, or the mean of all non-blank samples if needed).}

\item{filter_mz}{A boolean indicating whether or not to delete rows with masses ending in .8 or .9 (masses not found in natural products).}

\item{filter_rsd}{A boolean indicating whether or not to delete rows with too much relative standard deviation in each class.}

\item{filter_rsd_threshold}{A numerical threshold for relative standard deviation filtering: rows with relative standard deviation >= \code{filter_rsd_threshold} in each class are deleted. Only used if \code{filter_rsd} is \code{TRUE}.}
}
\value{
A list of 4 data.frames: \code{peak_data} contains the main data about detected peaks in positive and negative modes,
                                 \code{links} contains MSDial links between peaks,
                                 \code{detected_adducts} contains the probabilities of detection of adducts based on MSDial information.,
                                 \code{samples} contains samples information (names, classes, column names).
}
\description{
Import all data generated by MSDial.
}
