% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdial_functions.R
\name{import_msdial_file}
\alias{import_msdial_file}
\title{Import an MS file generated by MSDial.}
\usage{
import_msdial_file(source, samples)
}
\arguments{
\item{source}{A string indicating which mode to read from: "pos" or "neg".}

\item{samples}{A data.frame containing the samples information.}
}
\value{
A data.frame containing MSDial MS data.
}
\description{
Needs MSDial v4.00 or higher.
}
