% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfinder_functions.R
\name{import_msfinder_base_file}
\alias{import_msfinder_base_file}
\title{Generic function for importing a file generated by MSFinder}
\usage{
import_msfinder_base_file(filepath, columns_to_increment)
}
\arguments{
\item{filepath}{The path of the MSFinder file to import.}

\item{columns_to_increment}{A list of strings indicating which columns needs to have the final number in their name incremented.}
}
\value{
A data.frame containing MSFinder data.
}
\description{
Generic function for importing a file generated by MSFinder
}
