% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfinder_functions.R
\name{import_msfinder_data}
\alias{import_msfinder_data}
\title{Import files generated by MSFinder, both formula and structure at the same time.}
\usage{
import_msfinder_data(project_directory, source, level)
}
\arguments{
\item{project_directory}{The path of the project directory.}

\item{source}{A string indicating which mode to read from: "pos" or "neg".}

\item{level}{A string indicating the biosource level to consider.}
}
\value{
MSFinder formula and structure data joined together.
}
\description{
Import files generated by MSFinder, both formula and structure at the same time.
}
