% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfinder_functions.R
\name{import_msfinder_formula_file}
\alias{import_msfinder_formula_file}
\title{Import a formula file generated by MSFinder.}
\usage{
import_msfinder_formula_file(source, level)
}
\arguments{
\item{source}{A string indicating which mode to read from: "pos" or "neg".}

\item{level}{A string indicating the biosource level to consider.}
}
\value{
A data.frame containing MSFinder formula data.
}
\description{
Import a formula file generated by MSFinder.
}
