% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{increment_strings}
\alias{increment_strings}
\title{Increment the final number in strings.}
\usage{
increment_strings(strings_to_increment, patterns)
}
\arguments{
\item{strings_to_increment}{A list of strings to increment.}

\item{patterns}{A list of strings of patterns indicating which strings needs to be incremented.}
}
\value{
A list of strings identical to \code{strings_to_increment} but with strings names corresponding to \code{patterns} having had their final number incremented.
}
\description{
Increment the final number in strings.
}
