% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cad_pda_functions.R
\name{launch_cad_fusion}
\alias{launch_cad_fusion}
\title{Joins annotated MS data with CAD or PDA data.}
\usage{
launch_cad_fusion(project_directory, mode, threshold_rt)
}
\arguments{
\item{project_directory}{The path of the project directory.}

\item{mode}{"CAD" or "PDA".}

\item{threshold_rt}{A numerical value indicating the retention time tolerance.}
}
\description{
Joins annotated MS data with CAD or PDA data.
}
