% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_msfinder_annotation.R
\name{launch_msfinder_annotation}
\alias{launch_msfinder_annotation}
\title{Annotates peaks based on files extracted from MSFinder.}
\usage{
launch_msfinder_annotation(
  compound_levels = NULL,
  biosoc_levels = c("generic"),
  levels_scores = NULL
)
}
\arguments{
\item{compound_levels}{A vector containing the compound levels to consider, in the given order.}

\item{biosoc_levels}{A vector containing the biosource levels to consider, in the given order.}

\item{levels_scores}{A list of levels names and their corresponding multiplier to adapt final annotation scores.}
}
\description{
Annotates peaks based on files extracted from MSFinder.
}
\section{Architecture needed by launch_msfinder_annotation in the project_directory}{

\itemize{
  \item pos
  \itemize{
    \item msf_X
    \describe{
      \item{Formula<...>.txt}{Formulas for possible identifications in the X biosoc level, exported from MSFinder}
      \item{Structure<...>.txt}{Structures for possible identifications in the X biosoc level, exported from MSFinder}
    }
  }
  \item neg
  \itemize{
    \item msf_X
    \describe{
      \item{Formula<...>.txt}{Formulas for possible identifications in the X biosoc level, exported from MSFinder}
      \item{Structure<...>.txt}{Structures for possible identifications in the X biosoc level, exported from MSFinder}
    }
  }
  \item intermediary_data
  \describe{
    \item{samples.csv}{Information on samples}
    \item{final_peaks_data.csv}{Peaks remaining after all filters have been applied}
    \item{links_ms-final.csv}{Links between peaks indicated by MSDial (global) and adducts/neutral losses detection (by cluster)}
  }
}
}

\section{Output files of launch_msfinder_annotation in the project directory}{

\describe{
  \item{annotated_data-cleaned.csv}{Final auto-annotated data with only relevant peaks}
}
\itemize{
  \item intermediary_data
  \describe{
    \item{identifying_data.csv}{Union of MSDial and MSFinder data, with clusters info}
    \item{annotated_data.csv}{Final auto-annotated data, complete with all possibilities}
    \item{annotated_data-manual_check.csv}{Final auto-annotated data for clusters needing manual checking (only created in this case)}
  }
}
}

