% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscleanr.R
\docType{data}
\name{mass_adducts_neg}
\alias{mass_adducts_neg}
\title{Mass differences of adducts with the original compound M, in negative mode.}
\format{A data frame with 29 rows and 2 variables:
\describe{
  \item{adduct}{name of the adduct}
  \item{diff}{mass difference with M in negative mode}
}}
\usage{
mass_adducts_neg
}
\description{
Mass differences of adducts with the original compound M, in negative mode.
}
\seealso{
\code{\link{mass_adducts_pos}}, \code{\link{mass_neutral_loss_neg}}, \code{\link{mass_neutral_loss_pos}}, \code{\link{mass_isotopes}}
}
\keyword{datasets}
