% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscleanr.R
\docType{data}
\name{mass_neutral_loss_neg}
\alias{mass_neutral_loss_neg}
\title{Masses of neutral losses, in negative mode.}
\format{A data frame with 2 rows and 2 variables:
\describe{
  \item{name}{name of the neutral loss}
  \item{m}{mass in negative mode}
}}
\usage{
mass_neutral_loss_neg
}
\description{
Masses of neutral losses, in negative mode.
}
\keyword{datasets}
