% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscleanr.R
\docType{data}
\name{mass_neutral_loss_pos}
\alias{mass_neutral_loss_pos}
\title{Masses of neutral losses, in positive mode.}
\format{A data frame with 2 rows and 2 variables:
\describe{
  \item{name}{name of the neutral loss}
  \item{m}{mass in positive mode}
}}
\usage{
mass_neutral_loss_pos
}
\description{
Masses of neutral losses, in positive mode.
}
\seealso{
\code{\link{mass_adducts_neg}}, \code{\link{mass_adducts_pos}}, \code{\link{mass_neutral_loss_neg}}, \code{\link{mass_isotopes}}
}
\keyword{datasets}
