% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{Transform a matrix to a data.frame.}
\usage{
matrix_to_df(matrix_in, value_name = "Freq")
}
\arguments{
\item{matrix_in}{A matrix to transform.}

\item{value_name}{A string indicating how to name the value column.}
}
\value{
A 3-columns data.frame
}
\description{
Transform a matrix to a data.frame.
}
