% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscleanr.R
\docType{package}
\name{mscleanr}
\alias{mscleanr}
\title{mscleanr: A package for cleaning and analyzing MS data}
\description{
The mscleanr package provides 2 important functions: \code{\link{clean_msdial_data}}, \code{\link{keep_top_peaks}} and \code{\link{launch_msfinder_annotation}}.
See the functions documentation and vignettes for more information.
Needs MSDial v4.00 or higher.
}
\section{Architecture needed for an analysis}{

See functions documentation.
}

