% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print_peaks_status}
\alias{print_peaks_status}
\title{Print number of peaks in data in a message for the user.}
\usage{
print_peaks_status(peaks, msg)
}
\arguments{
\item{peaks}{A data.frame contaning peaks data.}

\item{msg}{A string containing the message content.}
}
\description{
Print number of peaks in data in a message for the user.
}
