% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_help.R
\name{recurrent_params}
\alias{recurrent_params}
\title{Return documentation info for recurrent parameters.}
\usage{
recurrent_params(...)
}
\arguments{
\item{...}{A list of parameters whose documentation is needed.}
}
\description{
Return documentation info for recurrent parameters.
}
