% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{stop_script}
\alias{stop_script}
\title{Stop the script after printing an error.}
\usage{
stop_script(..., generic_msg = TRUE)
}
\arguments{
\item{...}{Strings to print.}

\item{generic_msg}{Print the generic help message after the error.}
}
\description{
Stop the script after printing an error.
}
