% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{substrRight}
\alias{substrRight}
\title{Extract characters from a string from the right.}
\usage{
substrRight(x, n = 1)
}
\arguments{
\item{x}{A string.}

\item{n}{An integer with the number of characters to get.}
}
\value{
A string of length n.
}
\description{
Extract characters from a string from the right.
}
