% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{update_annotations_scores}
\alias{update_annotations_scores}
\title{Adapt the final annotations scores.}
\usage{
update_annotations_scores(data, focus_level, coef_mult)
}
\arguments{
\item{data}{A data.frame containing annotations.}

\item{focus_level}{A string indicating the biosource or compound level to consider.}

\item{coef_mult}{A numeric value indicating the coefficient by which to multiply the original MSFinder score.}
}
\value{
A data.frame containing annotations with multiplied scores.
}
\description{
Adapt the final annotations scores.
}
