% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_architecture_for_clean_msdial_data}
\alias{check_architecture_for_clean_msdial_data}
\title{Check project architecture before running \code{\link{clean_msdial_data}}.}
\usage{
check_architecture_for_clean_msdial_data(project_directory, overwrite = FALSE)
}
\arguments{
\item{project_directory}{The path of the project directory.}

\item{overwrite}{A boolean indicating whether to stop the script if there is an existing analysis or overwrite it.}
}
\description{
Check project architecture before running \code{\link{clean_msdial_data}}.
}
