% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_architecture_for_launch_msfinder_annotation}
\alias{check_architecture_for_launch_msfinder_annotation}
\title{Check project architecture before running \code{\link{launch_msfinder_annotation}}.}
\usage{
check_architecture_for_launch_msfinder_annotation(
  msfinder_biosoc_levels = NULL
)
}
\arguments{
\item{msfinder_biosoc_levels}{MSFinder annotation biosource levels to check (only used when \code{launch_msfinder_annotation = TRUE}).}
}
\description{
Check project architecture before running \code{\link{launch_msfinder_annotation}}.
}
