% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_main_architecture}
\alias{check_main_architecture}
\title{Check the project main architecture}
\usage{
check_main_architecture(project_directory)
}
\arguments{
\item{project_directory}{The path of the project directory.}
}
\description{
Check the project main architecture
}
