% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_csv_to_msp.R
\name{convert_csv_to_msp}
\alias{convert_csv_to_msp}
\title{Convert the final CSV file post annotations to MSP format.}
\usage{
convert_csv_to_msp(project_directory, min_score = 20)
}
\arguments{
\item{project_directory}{The path of the project directory.}

\item{min_score}{Peaks must have a final score >= \code{min_score} to be exported to the msp file.}
}
\description{
Convert the final CSV file post annotations to MSP format.
}
