% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_outputs.R
\name{export_params}
\alias{export_params}
\title{Export parameters used for the current run.}
\usage{
export_params(project_directory, ...)
}
\arguments{
\item{project_directory}{The path of the project directory.}

\item{...}{Parameters used for the current run of the script.}
}
\description{
Export parameters used for the current run.
}
