% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_top_peaks.R
\name{keep_top_peaks}
\alias{keep_top_peaks}
\title{Filter peaks based on a given criterion.}
\usage{
keep_top_peaks(project_directory, selection_criterion, n = 1,
  export_filtered_peaks = TRUE, overwrite = FALSE)
}
\arguments{
\item{project_directory}{The path of the project directory.}

\item{selection_criterion}{"intensity" or "degree" or "both".}

\item{n}{Keep only n.top peaks by clusters if intensity or degree filters.}

\item{export_filtered_peaks}{Copy peaks files corresponding to the peaks remaining after filtering in a new folder.}

\item{overwrite}{A boolean indicating whether to stop the script if there is an existing analysis or overwrite it.}
}
\description{
Keeps \code{n} peaks by cluster.
If there are ties, more than \code{n} peaks can be selected ( see \code{\link[dplyr]{top_n}} for more information).
}
