% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscleanr.R
\docType{data}
\name{mass_isotopes}
\alias{mass_isotopes}
\title{Masses of isotopes.}
\format{A data frame with 2 rows and 2 variables:
\describe{
  \item{name}{name of the isotope}
  \item{m}{mass}
}}
\usage{
mass_isotopes
}
\description{
Masses of isotopes.
}
\keyword{datasets}
